Almost Done

+++ forms/employeeDeleter/employeeDeleter.js
@@ -1,24 +1,38 @@
-
-
 rbtnDelete.onchange = function() {
-    if (typeof(s) == "object") {
-        return
-    } else {
-        let userChoice = $("input[name=rbtnDelete]:checked").prop("value")
+        if (typeof(s) == "object") {
+            return
+        } else {
+            let userChoice = $("input[name=rbtnDelete]:checked").prop("value")
+            let found = false
 
-        query = `DELETE FROM customer WHERE name = '${userChoice}'`
-        console.log(query)
-        req = Ajax("https://ormond.creighton.edu/courses/375/ajax-connection.php", "POST", "host=ormond.creighton.edu&user=" + netID + "&pass=" + pw + "&database=" + netID + "&query=" + query)
+            query = "SELECT name FROM customer"
+            req = Ajax("https://ormond.creighton.edu/courses/375/ajax-connection.php", "POST", "host=ormond.creighton.edu&user=" + netID + "&pass=" + pw + "&database=" + netID + "&query=" + query)
 
-        if (req.status == 200) //transit worked.
-            if (req.responseText == 500)
-                lblMessage2.textContent = `You have successfully deleted the employee named ${userChoice}`
-            else
-                lblMessage2.textContent = `There was a problem deleting ${userChoice} from the database.`
-        else
-            lblMessage2.textContent = `Error: ${req.status}`
-    }
-}
+            if (req.status == 200)
+                results = JSON.parse(req.responseText)
+
+            for (i = 0; i < results.length; i++) {
+                if (userChoice == results[i][1]) {
+                    found = true
+                    break
+                }
+            }
+            if (found == false)
+                lblMessage2.value = "That pet name is not in the database."
+            else {
+                query = `DELETE FROM customer WHERE name = '${userChoice}'`
+                console.log(query)
+                req = Ajax("https://ormond.creighton.edu/courses/375/ajax-connection.php", "POST", "host=ormond.creighton.edu&user=" + netID + "&pass=" + pw + "&database=" + netID + "&query=" + query)
+
+                if (req.status == 200) //transit worked.
+                    if (req.responseText == 500)
+                        lblMessage2.textContent = `You have successfully deleted the employee named ${userChoice}`
+                else
+                    lblMessage2.textContent = `There was a problem deleting ${userChoice} from the database.`
+                else
+                    lblMessage2.textContent = `Error: ${req.status}`
+            }
+        }
 
 btnDisplay.onclick = function() {
     rbtnDelete.clear()
@@ -26,7 +40,6 @@ btnDisplay.onclick = function() {
     rbtnDelete.items = ''
     query = "SELECT name FROM customer"
     req = Ajax("https://ormond.creighton.edu/courses/375/ajax-connection.php", "POST", "host=ormond.creighton.edu&user=" + netID + "&pass=" + pw + "&database=" + netID + "&query=" + query)
-
     if (req.status == 200) { //transit trip worked. 
         results = JSON.parse(req.responseText)
         if (results.length == 0)
@@ -36,6 +49,7 @@ btnDisplay.onclick = function() {
                 rbtnDelete.addItem(results[i])
         }
 
-    } else // the transit didn't work - bad wifi? server turned off?
+    } 
+    else // the transit didn't work - bad wifi? server turned off?
         lblMessage2.value = "Error code: " + req.status
 }
